requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4
django-ckeditor==6.4.0
django-js-asset==2.0.0
django-phonenumber-field==6.1.0
django-taggit==3.0.0
phonenumbers==8.12.47
Pillow==9.1.0
sqlparse==0.4.2
